refactor: update namespaces and user ID types
Changed the namespace for multiple classes from `CoolMeals.DAL.Models` to `CoolMeals.DAL.Database.Models` across various files to improve organization. Updated the `UserId` type in the `ShoppingList`, `UserDiet`, and `UserIntolerance` classes from `int` to `Guid` for better user identification. Reintroduced the `User` property in the `Pantry` and `Recipe` classes, enhancing the relationship with the `ApplicationUser` class. These changes aim to streamline the codebase and enhance user data handling.
