+++ src/CoolMeals.DAL/Database/Models/ApplicationUser.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Identity;
 
-namespace CoolMeals.DAL.Models;
+namespace CoolMeals.DAL.Database.Models;
 public class ApplicationUser : IdentityUser<Guid>
 {
     public string FirstName { get; set; } = string.Empty;
