+++ src/CoolMeals.DAL/Database/Models/Diet.cs
@@ -1,4 +1,4 @@
-﻿namespace CoolMeals.DAL.Models;
+﻿namespace CoolMeals.DAL.Database.Models;
 public class Diet
 {
     public int DietId { get; set; }
