+++ src/CoolMeals.DAL/Database/Models/Recipe.cs
@@ -1,4 +1,4 @@
-﻿namespace CoolMeals.DAL.Models;
+﻿namespace CoolMeals.DAL.Database.Models;
 public class Recipe
 {
     public int RecipeId { get; set; }
@@ -9,8 +9,8 @@ public class Recipe
     public int? Servings { get; set; }
     public string? ImageUrl { get; set; }
     public bool IsFavorite { get; set; }
-    public ApplicationUser User { get; set; } = null!;
     public ICollection<RecipeIngredient> RecipeIngredients { get; } = new List<RecipeIngredient>();
     public ICollection<Instruction> Instructions { get; } = new List<Instruction>();
     public ICollection<MealPlanEntry> MealPlanEntries { get; } = new List<MealPlanEntry>();
+    public ApplicationUser User { get; set; } = null!;
 }
